requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
torchvision>=0.10.0
numpy>=1.19.2
scipy>=1.7.0
librosa>=0.8.1
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
opencv-python>=4.5.3
PyYAML>=5.4.1
tqdm>=4.62.3
pandas>=1.3.3
pillow>=8.3.2
soundfile>=0.10.3
tensorboard>=2.6.0
einops>=0.3.2  # for Vision Transformers
timm>=0.4.12   # for EfficientNet
